actualitzacio

+++ P0-FirstContact/P99182.cc
@@ -1,10 +1,14 @@
-// media entre dos numeros
+// Promedio
+// Escribe el promedio de dos números enteros.
+// input: Dos numeros de -1000 y 1000, separados por un espacio.
+// output: Escribe una línea con el promedio de ambos números.
 #include <iostream>
 using namespace std;
 
 int main()
 {
-    double n1, n2;
+    int n1, n2;
     cin >> n1 >> n2;
-    cout << (n1 + n2) / 2 << endl;
+
+    cout << (n1 + n2) / 2.0 << endl;
 }
